requirements.txt
flask
python-dotenv
aiohttp
discord.py
